serverutils: nil check system config cache
See [here](https://teamcity.cockroachdb.com/buildConfiguration/Cockroach_BazelEssentialCi/6797223?showRootCauses=false&expandBuildChangesSection=true&expandBuildDeploymentsSection=true&expandBuildProblemsSection=true&expandBuildTestsSection=true)

Release note: None

+++ pkg/testutils/serverutils/test_cluster_shim.go
@@ -263,8 +263,9 @@ func StartNewTestCluster(
 	cluster.Start(t)
 	for i := 0; i < cluster.NumServers(); i++ {
 		sysconfigProvider := cluster.Server(i).SystemConfigProvider()
-		sysconfig := sysconfigProvider.GetSystemConfig()
-		sysconfig.PurgeZoneConfigCache()
+		if sysconfig := sysconfigProvider.GetSystemConfig(); sysconfig != nil {
+			sysconfig.PurgeZoneConfigCache()
+		}
 	}
 	return cluster
 }
